Use 'as' when catching exceptions instead of ','.
Since we are already using Python 2.6, this doesn't interfere with our
ability to be executed on earlier versions of Python.

This is also one little step closer to being compatible with Python 3. This
may or may never happen, who knows, as there are many other changes, but the
goal is to have code that is easier to read and maintain.