Add initial configuration for travis-ci.
This is only part of what is needed for the continuous integration, but it
is able to run our basic unittests.

This does encourage us to create more unittests.

Another side benefit is that once we have a good coverage of unittests, the
pull requests that we receive will (seemingly) be tested by travis and we
will know if such pull requests are likely to create regressions.