Remove passing of unneeded arguments.
We are passing the name of a file with cookies to functions that simply
ignore cookies.

In particular, we are creating files with cookies, saving cookie jars and
loading cookie jars all over the place and this is not smart.

We should only load/store cookies if we receive a (debug-only) option from
the user to store those cookies to files.