Convert float to int for Python 3 compatibility.
When running the script with `python2.7 -3`, we get many warnings about
deprecation comparing different types is one of those. So, we coerce the
type returned by mtime to be an int.

The reason is that we don't actually need precision float precision and it
may give the interpreter a little bit of an easier time translating stuff.