Fix the real reason for warning of comparison of different types.
The real reason why we get a warning that we are comparing multiple
different types is that last_update is initialized with Null and, then, used
in the computation of a maximum.

Since we are dealing with seconds since the epoch, we can assume that
everything that is a time is non-negative and we set the initial "minus
infinity" to be -1.  If needed, we can put the minimum value for an integer
as the "minus infinity".

This improves compatibility with Python 3.