Fix for #98 without breaking the runtest
For the wikipedia test, the parser fails to find mp4 files, which is correct.
However, if the parser fails to find mp4 files, it tries to look for hidden files, as for some of the University of Washington courses, and again fails, because there is no hidden mp4 file. 

I thus implemented a short test that checks if the hidden mp4 check finds mp4 files or not.