Open cookies file in universal newline mode
In the cookies files exported with the firefox addon on a mac, every
line is terminated with a carriage return \r. Loading this file into a
MozillaCookieJar has the unwanted effect of having every cookie value
ending with \r. With the universal newline mode enabled, the \r is
replaced with a \n.