Fix downloading when authenticated via cookies file
The global variable csrftoken was not set when a cookies file was used.
Hence downloading files did not work. Authentication is now split into
two stages:
- get the login cookies from www.coursera.org. If a cookies file is
used, it should already contain these cookies.
- get the session and csrf_token cookies from  class.coursera.org. If a
cookies file is used, we first check if we already have the cookies,
otherwise we retrieve them as usual.