Add downloaders module
This module provides the interfaces to the external downloaders (curl,
…) and our own native downloader.

The class structure is deliberately very simple:
A downloader is a class that extends `downloaders.Downloader`.
The only thing that it should implement is a `_start_download` method
that has as arguments the url to download and the target file.