Supress HTTPError in grab_hidden_video_url
If the script doesn't find a mp4 for a lecture, it checks if the
`data-modal-iframe` attribute is present and uses that as video link.
However, Coursera always generates those attributes, even if the
lecture doesn't have a video. Hence, we request an invalid url and get
a 404 HTTPError. To prevent this from ending the cript, we add a try
clause to the grab_hidden_video_url.