Fix downloading multiple resources with the same format
The resources of a lecture were stored in a dictionary, indexed by
their format. This meant that if a lecture contains more than one
resource with the same format (e.g. multiple pdf files), only the last
one gets downloaded. This commit fixes this problem.
We add the title of the resource to the filename of the downloaded
resource. However, for backward compatibility, we only do this for the
resources that we did not download because of the bug.