README: Add newer examples to avoid misleading people.
We had more than one issue filed where people thought that they only had to
specify the "base name" of the classes.

While this was true during the inception of Coursera, their naming of the
classes has changed during time and this makes the documentation clearer to
match the post-initial naming.