coursera: Add a resource filter.
This useful when courses offer videos in two resolutions (say, high
resolution and low resolution) and you only want one of them.

In my limited experience, the low resolution files are both *much* larger in
size (about 3 times the size of the respective regular resolution videos)
and uglier.

I avoid them with a regular expression passed as:

    --resource_filter '\d+_(?!Low.*Res).*'

Note that every character above in the negative lookup seem to matter.  If
the regular expression can be made shorter or easier to read, I'd like to
know.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>