Implement a basic, light sanitized version of filename sanitization.
This is potentially dangerous (we are external sources write almost
arbitrary strings as filenames), but some people may need filenames from
higher portions of the Unicode space.

This is still *not* hooked via a command line option, but is only for
evaluation purposes, to address issues pointed in comments of bug report

Still, the default will be to keep backwards compatibility and the user will
have to knowingly enable this feature.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>