coursera: Hook the restriction of filenames to the main program.
The implementation sucks right now and we should *sincerely* get rid passing
each attribute of the `args` object to functions. The better thing is
probably to simply pass `args` and let the functions use the parts of it
that they need.

Really, a function with more than a dozen parameters sucks, especially when
we are duplicating things that are already readily available.