travis: Hook `coverage` with `nosetests`.
We limit the package that we are interested in to only `coursera` (maybe to
be changed to `coursera_dl` after a potential merge with other projects?).

Given that the code tested is *not* 100% the same as the code executed
when really downloading things (due to stubs etc.), we need to check what
options we have to cover real execution as well.

Note that this still is not sufficient to get the results in
https://coveralls.io/r/coursera-dl/coursera as we need to communicate the
results to them.