requirements: Loosen the versions of the packages.
Some of the versions that we had fixed in the `requirements.txt` file are
old (see Issue #247) and they should serve only as a baseline for what `pip`
installs.

Furthermore, we should also keep an eye on how newer versions of our
dependencies work (via the automated tests), so as to assist our users
better.