coursera: Download subtitles for languages other than English.
This is an initial, brute-force, kludgy, not yet hooked implementation into
the main script, only tested interactively (read: "from ipython") for
downloading subtitles for other languages.

It is slow and should probably be replaced with the use of Coursera's API
(if it offers what we need) for a more site-friendly solution.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>