coursera: utils: Don't be so greedy removing remarks. Fixes #269.
A lot of instructors actually put remarks like "optional" or "advanced
material" or "skip if already familiar" inside parentheses.

The original regexp had the intention of removing trailing durations, but,
unfortunately, there is no standard used by the instructors and we ended up
removing precious metadata from the videos.

This patch is meant to fix this.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>