Fix issue#300
Somehow the order of cookies parameters are an important fact for
coursera!, for example if there are three parameters in cookies
called: 'csrf_token', 'maestro_login_flag' and 'CAUTH' in exactly that
order, coursera doesn't accept it and redirect the request to signin
page, but if without changing parametes values and just changing their
orders: 'maestor_login_flag', 'CAUTH' and 'csrf_token', coursera accepts
it and fetches the requested page.

Updated .gitignore file to ignore virtualenv folders: 'venv2' and
'venv3'.