Add Basic User Login Procedure
Added an authentication function which implements basic
user login procedure where client clicks `Signin` button
on coursera.org website and fills his/her Email and Password
to enter. In this procedure we use API Login v3 for authentication
and there is no need of `class_name` to fetch `csrftoken`.
This can be useful for future purposes, especially for downloading
`On-Demand` courses.