coursera: cookies: Restrict TLS to version 1.0.
After a considerable amount of RTFM, I try to summarize here what I
discovered, namely, that the problem is actually a conjunction of many
things:

* Python 2.6 and 2.7.x (with x <= 8) don't define the constants
  `ssl.PROTOCOL_TLSv1_1` and `ssl.PROTOCOL_TLSv1_2`. [They were introduced
  with Python 2.7.9][0]

* Having TLS v1.1 and TLS v1.2 seem to only be available with sufficiently
  new versions of OpenSSL. Again, this is reading from [the Python docs][0].

* urllib3, as a result, doesn't seem to support explicitly specifying which
  which version of TLS to use, as far as I can see it (both by reading the
  reading the code and by reading
  https://github.com/shazow/urllib3/issues/487#issuecomment-59382798).

[0]: https://docs.python.org/2/library/ssl.html?highlight=ssl#ssl.PROTOCOL_TLSv1

Therefore, I guess that the best that we can do so far is to only request
`ssl.PROTOCOL_TLSv1` to be used. While it seems that TLS v1.0 has issues, I
hope that the packages we depend on manage to find a way to negotiate higher
versions of TLS, so that we can change that accordingly in the future.

A not so good alternative to that would be to decide dynamically (and
monkeypatch) urllib3 to see which TLS protocols we have and what we can try
to convice it to be used.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>