setup.py: Fill in some data.
This is still not really satisfactory, as it seems that we have a limitation
in the number of authors (or it seems that we should use an ad-hoc format)
for specifying who the authors are, according to various sources, including:

https://docs.python.org/2/distutils/setupscript.html#additional-meta-data

For that reason, I refrained from telling that I am an (the?) author of this
program, as John should be equally credited here, if not more.  I only
listed myself as the maintainer of the program.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>