setup: Be more specific about Python versions supported.
I'm not really sure if we should declare compatibility with Python 3.2.
While we also test for pypy (via Travis CI), I am not really sure how to
specify that in the file (I still have to do some reading).

Also, I don't really expect any problems with the use of Python 3.5, but
since it is not (yet) tested with travis, I'm not yet putting it explicitly
here.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>