setup: Add explicit mention of the `coursera-dl` script.
Otherwise, it seems that the entry point script does not get included in the
source tarball.

I still don't know what the proper solution for Windows users would be
(e.g., the `coursera-dl.bat` file that we had contributed). Perhaps
including it also along with the `coursera-dl` script?

There should be a better way to do this, most likely.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>