README: Make pip installation the preferred one.
While we are at it, demote the "recommended" installation on unix systems to
simply be an alternative installation method.

This document should, now, be really shrunk, without all the justification
of why we use one module or another. This just wastes space, draws attention
from more important points and is not of interest to our final users.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>