fabfile: Use pandoc to generate README.rst.
The rendering of Markdown files in PyPI sucks so bad that I'm willing to go
the extra steps and convert our documentation to rst.

I am not (yet) converting our documentation rst, since I believe that
Markdown is much easier on my eyes (and I believe that it is more
lightweight too).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>