coursera/cookies: Try to work around import errors with requests/urrlib3.
Ubuntu 14.04's requests packages behave differently from the upstream
version of requests.

The upstream version of requests includes an embedded copy of urllib3 (and
other things, like its own trusted root certificates), but, by Debian's
guidelines, convenience copies of libraries are discouraged and generally
removed.

Current versions of Debian's requests packages take care of transparently
linking the separately downloaded requests with urllib3 with glue code that
makes them work as well as with upstream's version of requests.

Unfortunately, given some reports from users (see issue #331), the version
of requests that comes with Ubuntu 14.04 doesn't seem to have that glue code
working as we expect.

This commit tries to cope with that, by trying to import urllib3 from more
places to see if we can get farther with the script, instead of bombing with
an import error.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>