README: Minor formatting changes to help RST generation.
I'm not really sure if the problem with the (ugly) final generation of the
HTML is a limitation of pandoc's MD -> RST or if it is a limitation of
pandoc's RST -> HTML conversion, but these changes *try* to address some of
the problems to generate the README.rst file that PyPI expects.

Some of these changes are not really quite natural (they are artificially
"imposed" by markdown) and we have some trailing whitespace (which is
invisible, alas!) to allow for continuation of structures instead of
breaking them.

Nevertheless, at least *some* of this work would not be necessary if PyPI
were able to use vanilla markdown formatting for the README of projects.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>