setup.py: Make script work with Python 3.
Otherwise, we were generating lists as the output of the function when
reading `README.rst` and, in Python 3, we had an exception being thrown
saying that lists didn't have a `split` method.

This makes `read_file` always return a string, which makes it possible to
run:

        python3 setup.py sdist

which is a portability win, of course.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>