setup.py: Fix #354 by generating README.rst with setup.py.
This commit makes execution of setup.py generate the README.rst file by
calling pandoc.

We only stop execution of the program if README.md can not be converted into
a README.rst file, in which case we execute the show a message to the user.

I hope that, in the future, PyPI can accept Markdown files and that we can
stop having to convert between formats (and, also, get a more faithful
reproduction of the intended markup, since some of the conversions are,
unfortunately, lossy).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>