coursera/cookies: Add hint to use the --on-demand option.
We have seen a significant rise in people asking if the program supports
On-Demand courses (or even claiming that it doesn't work).

See issues #386, #380, #379, #333, among many others.

This ability was contributed by @MXAmin quite some time ago and we even have
it documented in our README file, right there at the top, in the "Features"
section.

Hopefully this makes it more explicit that we *do support* such courses.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>