README: Reword to say that requirements.txt is required.
While not strictly true, it will save me a lot of headaches compared to the
situation of people using very old dependencies.  Generally, it is preferred
to have newer versions than to have older dependencies.

I don't know if `pip` supports the equivalent of Ruby's Gemfiles `~=`
operator to specify versions that don't break API compatibility with the
version specified.

I saw that mentioned in [PEP-400][pep400], but I don't know if it is
implemented already.

[pep440]: https://www.python.org/dev/peps/pep-0440/#compatible-release

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>