Deprecate --on-demand option, try to detect class type automatically
We have received a lot of issues where people are confused about
distinction between old-style and new-style (on-demand) courses.
Sometimes --on-demand option is required, sometimes it's not.

This patch deprecates --on-demand option in favour of automatic
detection. The detection is straight forward: first, try to treat
the course as an old-style, and if the course is not found, try to
treat it as a new-style (on-demand).

Corresponding messages were adjusted (login method). Help message
for --on-demand option now advices to remove that option.