README.md: Fix URL to codeclimate.
Unfortunately, with this move, we lost the history of the project.

On the other hand, the site doesn't seem to have any easy to spot place to
update the URL (the one that it has is not editable).

At least we now point the site to the correct repository. The previously
computed "GPA" was only 1.67 and we have a substantial amount of new tests
now.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>