README.md: Fix badge type from coveralls.io.
We use an SVG file instead of a PNG file, which should look better on some
devices (yes, I know, this is a frivolous change).

Unfortunately, the URL registered there is still `coursera-dl/coursera` and
not the new one (`coursera-dl/coursera-dl`), since the site gives me an HTTP
500 error whenever I click on the button to update the repository
information, and this seems to be the recommended update procedure.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>