Extract additional another variation of programming assignments
There are courses (e.g. data-structures, algorithms-on-graphs) that
have almost empty modules in api/opencourse.v1/course response. Such
modules look as follows:

{ // this is a module
    "id": "QgCuM", // <- lecture id
    "slug": "programming-assignment",
    "name": "Programming Assignment",
    "timeCommitment": 0,
    "elements": []
}

Their elements are empty so it's not straightforward to extract
supplementary materials (such as zip archives, PDFs) from them.
To do that, one needs to call api/onDemandCourseMaterials.v1
and parse the result. There you will see something like this:

{
  "id": "AUd0k",
  "moduleId": "0MGvs",
  "lessonId": "QgCuM",
  "name": "Programming Assignment 1: Decomposition of Graphs",
  "slug": "programming-assignment-1-decomposition-of-graphs",
  "timeCommitment": 10800000,
  "content": {
    "typeName": "gradedProgramming",
    "definition": {
      "programmingAssignmentId": "zHzR5yhHEeaE0BKOcl4zJQ@2",
      "gradingWeight": 20
    }
  },
  "isLocked": true,
  "itemLockedReasonCode": "PREMIUM",
  "trackId": "core"
},

Result is substituted into the list of lectures and is parsed
as a normal lecture.