Fix External Downloaders / Logging / Local Page failure / Wait Interval
There are really several changes in this commit, which are distinct but
hard to separate at this point.  Apologies.

1) Fix Aria2 (and probably other external downloaders), that break with
the "\\?\" that was added for Windows.  This took a bit of refactoring
-- the call to normalize_path() converts a relative path to an absolute
path, and we need the same sort of path no matter what downloader we
use.  So I moved the call to this function up a level, so that I could
make the path absolute no matter what, and then only "normalize" it if
we were using the NativeDownloader.  It then has to be called from both
the "old style" and "on demand" download methods.

The external downloaders still fail on long filenames, but at least they
work most of the time (rather than not at all).  We need another
solution for the long file names, or, we can do most of the downloading
with the external one of choice, and then do a final run with the
native(/internal) downloader, to get the missed files.

2)  The arguments that were being passed to Aria2 (1.24.0) were wrong.
Notably, it needed to have the target path and filename separated, which
is why I still needed the absolute path as above.

3) The logging of skipped and failed URLs is enhanced to also write the
results to a timestampped file, as well as stdout at the end of
execution.  This allows the user to reference these files later, when
they come back to the class.

3a) In order to track the failed URLs, the success/failure reporting had
to be normalized up the call stack.  The lower level functions were
returning just a "last update" for the downloaded file, or the external
downloaders were just returning a process exit code; but the higher
level ones were returning success/failure.  So I added the
success/failure, and also refactored the failure container to always
hold these URLs, and convey them up the stack.

3b) To both the URL success and failure logs, I added the file the
program was trying to write, in addition to the URL it was trying to
download.  This way, users can tell where the skipped/failed URL goes in
the class syllabus, and/or manually download it if they wish.

4) The "local page" option was failing because it was writing
(apparently) binary "UTF-8", but that wasn't specified in the open
statements.  Simply adding "wb" or "rb" fixed the problem.

5) I changed the native downloader's "wait_interval" pause factor (base
of the exponentiation) from 2 to 1.5.  There were complaints that the 63
second pause was annoyingly long, and I don't see its value.  This still
gives a decent pause between retries, but doesn't wait so long.  If
there's some reason for the longer pause, it can be changed back easily
enough.