coursera/downloaders: Add filename shortening.
This is wildly experimental, a work-in-progress, has hardcoded magic numbers
and generally sucks.

But it seems to be reasonable (in spirit, at least) in face of all the
problems with Windows having problems with long filenames that we don't
experience in a Unix system.

Once again, this is just a proposal and I don't really expect it to be
merged, unless there is some cleaning up (like extracting the magic
constants etc.).