First take on implementing parallel downloader
The idea is to separate selection of resources to download and
actual downloading. Moreover, it should be a transparent
replacement of sequential downloader for parallel downloader.

On of the issues to keep in mind when offloading download
process into a separate thread is how to pass exceptions
back to the main thread.