Add tox to requirements.txt and extend it to versions 26,27,33,34,35
Now it's a snap to run tests locally on multiple versions of Python.

Also fix install_command. tox used to fail because it tried to
pass `--download-cache` option to pip, which is obsolete and gone.

Update CONTRIBUTING.md to explain how to use tox and pyenv together
to run tests locally on multiple Python versions.