coursera-dl: Use a Python 3 shebang.
This makes it more convenient for when you have your modules installed from
your Linux distribution or so.

As it only affects those that are working with a cloned git tree, it is
supposed to be the case the they know what they are doing and shouldn't
affect the vast majority of the users (the wrapper created by setup.py is a
completely different one, when users install via pip etc.).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>